==> data_processing/5a-join_mobility_data_generate_alters.ipynb <==
library(readr)
library(dplyr)
library(tidyr)
library(stringr)
library(doMC)
library(lubridate)
library(USAboundaries)
library(sf)
library(blscrapeR)
registerDoMC(4)

fp      <- '/pool001/mfzhao/'
w2020   <- read_csv(str_c(fp, 'PROCESSED_DATA/region_weather_2020.csv'))
w2019   <- read_csv(str_c(fp, 'PROCESSED_DATA/region_weather_2020.csv'))
policy  <- read_csv(str_c(fp, 'PROCESSED_DATA/policy.csv'))
sci     <- read_delim(str_c(fp, 'sci/SCI_county.tsv'),  "\t", escape_double = FALSE, trim_ws = TRUE)
nyt     <- read_csv(str_c(fp, 'nyt_covid/us-counties.csv'))
us_pop  <- read_csv('/pool001/mfzhao/geo_data/cc-est2018-alldata.csv')
usm2019 <- read_csv(str_c(fp, 'safegraph/us_mobility_2019.csv'))
usm2020 <- read_csv(str_c(fp, 'safegraph/us_mobility.csv'), 
                    col_types = cols(
                        mhdt = col_double(),
                        mnhdt = col_double(),
                        mdtfh = col_double()))

us_pop %>%
    unite('key', STATE, COUNTY, sep = '') %>%
    filter(YEAR == 11, 
           AGEGRP==0) %>% 
    group_by(key) %>%
    summarize(n = sum(TOT_POP)) -> us_pop

usm2020 %>%
    rename(key = origin_county) %>%
    filter(date >= as.Date('2020-01-01'), date < as.Date('2020-07-01')) -> usm2020

usm2020 %>%
    filter(key < '57000') %>%
    group_by(key) %>%
    summarize(mdc = mean(device_count), n = n()) %>%
    filter(n == max(n), mdc > 500)%>%
    select(key) %>%
    left_join(usm2020) -> usm2020

usm2020 %>%
    group_by(key) %>%
    tally() %>%
    filter(n == max(n)) %>%
    select(key) -> keys_mobility

policy %>%
    select(key) %>%
    distinct() -> keys_policy

w2020 %>%
    select(key) %>%
    distinct() %>%
    anti_join(w2020 %>% 
              filter(is.na(PRCP) | is.na(TMAX)) %>%
              select(key) %>%
              distinct()) -> keys_weather

sci %>%
    select(key = user_county) %>%
    distinct() -> keys_sci

keys_mobility %>%
    inner_join(keys_policy) %>%
    inner_join(keys_weather) %>%
    inner_join(keys_sci) -> keys

expand.grid(key = keys$key, date = seq.Date(as.Date('2020-01-01'), as.Date('2020-06-30'), '1 day'), 
            stringsAsFactors = F) %>%
    left_join(nyt %>% rename(key = fips)) %>%
    fill(cases, deaths) %>%
    replace_na(list(cases = 0, deaths = 0)) %>%
    select(-county, -state) %>%
    group_by(key) %>%
    arrange(key, date) %>%
    mutate(newcases  = cases - lag(cases),
           newdeaths = deaths - lag(deaths),
           newcases  = ifelse(newcases < 0, 0, newcases),
           newdeaths = ifelse(newdeaths < 0, 0, newdeaths)) %>%
    replace_na(list(newcases = 0, newdeaths = 0)) -> nyt

keys %>%
    left_join(us_pop) -> population

write_csv(keys, str_c(fp, 'PROCESSED_DATA/keys.csv'))

county_sf <- us_counties()

as.data.frame(county_sf) %>%
    select(-geometry) %>%
    mutate(L3 = row_number(),
           key = str_c(statefp, countyfp)) %>%
    left_join(as.data.frame(st_coordinates(county_sf))) %>%
    select(key, X, Y) -> coords

coords %>%
    inner_join(coords, by = c('X', 'Y')) %>%
    filter(key.x != key.y) %>%
    select(origin_county = key.x,
           destination_county = key.y) %>%
    distinct() %>%
    mutate(bordering = 1) %>%
    arrange(origin_county, destination_county) -> bordering

county_sf %>%
    st_centroid() %>%
    as.data.frame() %>%
    mutate(key = str_c(statefp, countyfp)) %>%
    select(key, geometry) %>%
    arrange(key) -> centroids

distm <- data.frame(st_distance(centroids$geometry, centroids$geometry))
colnames(distm) <- centroids$key

distm %>%
    mutate(origin_county = centroids$key) %>%
    gather(key = 'destination_county', value = 'dist', -origin_county) %>%
    mutate(dist = as.numeric(dist)/1000) %>% 
    inner_join(keys, by = c('origin_county' = 'key')) %>%
    inner_join(keys, by = c('destination_county' = 'key')) -> distm

distm %>%
    mutate(nearby = as.numeric(dist < 100)) %>%
    select(-dist) %>%
    spread(key = destination_county, value = nearby) %>%
    select(-origin_county) %>%
    as.matrix() -> near

sci %>%
    inner_join(keys, by = c('user_county' = 'key')) %>%
    inner_join(keys, by = c('fr_county' = 'key'))  %>%
    inner_join(population, by = c('fr_county' = 'key')) %>%
    arrange(user_county, fr_county) -> sci

write_csv(sci, str_c(fp, 'PROCESSED_DATA/processed_sci.csv'))

sci %>%
    mutate(user_state = str_sub(user_county, 1, 2),
           fr_state   = str_sub(fr_county, 1, 2)) %>%
    group_by(user_county) %>%
    mutate(w = ifelse(user_state == fr_state, 0, scaled_sci * n),
           w = w/sum(w)) %>%
    select(key = user_county, fr_county, w) %>%
    spread(key = fr_county, value = w) %>%
    ungroup(key) %>%
    arrange(key) %>%
    select(-key) %>%
    as.matrix() -> stateWM

rownames(stateWM) <- keys$key
write_rds(stateWM, str_c(fp, 'PROCESSED_DATA/stateWM.RDS'))

sci %>%
    group_by(user_county) %>%
    mutate(w = ifelse(user_county == fr_county, 0, scaled_sci * n),
           w = w/sum(w)) %>%
    select(key = user_county, fr_county, w) %>%
    spread(key = fr_county, value = w) %>%
    ungroup(key) %>%
    arrange(key) %>%
    select(-key) %>%
    as.matrix() -> WM

rownames(WM) <- keys$key
write_rds(WM, str_c(fp, 'PROCESSED_DATA/WM.RDS'))

stateWM_near <- stateWM * near 
stateWM_far  <- stateWM * (1 - near)
write_rds(stateWM_near, str_c(fp, 'PROCESSED_DATA/stateWM_near.RDS'))
write_rds(stateWM_far, str_c(fp, 'PROCESSED_DATA/stateWM_far.RDS'))

sci %>%
    mutate(user_state = str_sub(user_county, 1, 2),
           fr_state   = str_sub(fr_county, 1, 2)) %>%
    filter(user_county != fr_county) %>%
    group_by(user_county) %>%
    mutate(w = scaled_sci * n,
           w = ifelse(user_state == fr_state, 0, w)/sum(w)) %>%
    select(key = user_county, fr_county, w) %>%
    spread(key = fr_county, value = w) %>%
    ungroup(key) %>%
    arrange(key) %>%
    select(-key) %>%
    as.matrix() %>%
    replace(is.na(.), 0) -> stateWM_v2

write_rds(stateWM_v2, str_c(fp, 'PROCESSED_DATA/stateWM_v2.RDS'))

sci %>%
    mutate(w = scaled_sci * n,
           diff_state = as.numeric(str_sub(user_county, 1, 2) != str_sub(fr_county, 1, 2)),
           diff_county = as.numeric(user_county != fr_county)) %>%
    group_by(user_county) %>%
    summarize(all = sum(w), 
              diff_county = sum(w * diff_county),
              diff_state = sum(w * diff_state)) %>%
    mutate(dcp = diff_county/all,
           dsp = diff_state/all) %>%
    left_join(us_pop, by = c('user_county' = 'key')) %>%
    summarize(dcp = sum(dcp * n/sum(n)),
              dsp = sum(dsp * n/sum(n))) %>%
    mutate(scp = 1 - dcp)

usm2019 %>%
    select(key = origin_county,
           date,
           pchd.2019 = pchd,
           mcbgv.2019 = mcbgv,
           plt1hafh.2019 = plt1hafh,
           ppthgt75.2019 = ppthgt75,
           plt2kmt.2019  = plt2kmt) %>%
    mutate(month = month(date),
           day   = day(date)) %>%
    select(-date) -> usm2019

w2020 %>%
    mutate(month = month(DATE),
           day   = date(DATE)) %>%
    left_join(w2019 %>% 
              mutate(month = month(DATE),
                     day   = date(DATE)) %>%
              select(key, month, day, PRCP.2019 = PRCP, TMAX.2019 = TMAX)) %>%
    select(key,
           date = DATE,
           PRCP,
           TMAX,
           PRCP.2019,
           TMAX.2019) -> weather

usm2020 %>%
    inner_join(keys) %>%
    mutate(month = month(date),
           day   = day(date)) %>%
    select(-mhdt, -mnhdt, -mdtfh, -device_count) %>%
    left_join(usm2019) %>%
    mutate(pnchd      = 1 - pchd,
           pgt1hafh   = 1 - plt1hafh,
           ppthlt75   = 1 - ppthgt75,
           pgt2kmt    = 1 - plt2kmt,
           dmcbgv     = mcbgv - mcbgv.2019,
           dpnchd     = pchd.2019 - pchd,
           dpgt1hafh  = plt1hafh.2019 - plt1hafh,
           dppthlt75  = ppthgt75.2019 - ppthgt75,
           dpgt2kmt   = plt2kmt.2019 - plt2kmt,
           pcmcbgv    = dmcbgv/mcbgv.2019,
           pcpnchd    = dpnchd/(1 - pchd.2019),
           pcpgt1hafh = dpgt1hafh/(1 - plt1hafh.2019),
           pcppthlt75 = dppthlt75/(1 - ppthgt75.2019),
           pcpgt2kmt  = dpgt2kmt/(1 - plt2kmt.2019),
           cluster    = str_sub(key, 1, 2)) %>%
    select(-matches('2019'), 
           -month, 
           -day,
           -pchd,
           -plt1hafh, 
           -ppthgt75,
           -plt2kmt) %>%
    left_join(us_pop) %>%
    left_join(nyt) %>%
    inner_join(policy) %>%
    inner_join(weather) %>%
    arrange(date, key) -> df

weightedAlters <- function(df, wm, ...) {
    df %>% 
        select(date, key, ...) %>%
        spread(key = key, value = ...) %>%
        ungroup() %>%
        select(-date) %>%
        as.matrix() -> txn_data
    
    df %>%
        ungroup() %>%
        select(date) %>%
        distinct() %>%
        arrange(date) -> dates
    
    outMatrix <- tcrossprod(txn_data, wm)
    colnames(outMatrix) <- colnames(txn_data)
    
    data.frame(dates, outMatrix) %>%
        gather(key = 'key', value = 'value', -date) %>%
        arrange(date, key) %>%
        select(-date, -key) -> out_df
    return(out_df$value)
}

nquantiles <- 10

df %>%
    ungroup() %>%
    select(date, key, PRCP, TMAX) %>%
    mutate(PRCP = ifelse(PRCP == 0, NA, PRCP),
           pq = ifelse(is.na(PRCP), 0, ntile(PRCP, nquantiles)), 
           tq = ntile(TMAX, nquantiles)) -> weather_quantiles

wq_ind <- function(thres) {
    weather_quantiles %>%
        transmute(tempname1 = as.numeric(pq >= thres),
                  tempname2 = as.numeric(tq >= thres)) -> temp
    colnames(temp) <- c(str_c('prcp', str_pad(thres, 2, pad = '0')), str_c('tmax', str_pad(thres, 2, pad = '0'))) 
    return(temp)
}

wq_inds <- foreach(i = 1:nquantiles, .combine = cbind) %dopar% wq_ind(i)
wq_colnames <- c(str_c('prcp', str_pad(1:nquantiles, 2, pad = '0')), str_c('tmax', str_pad(1:nquantiles, 2, pad = '0')))

df %>%
    bind_cols(wq_inds[wq_colnames] %>% select(-tmax01)) -> df

colnames(df)

df %>%
    mutate_at(vars(mcbgv, pnchd, pgt1hafh, ppthlt75, pgt2kmt, 
                   dmcbgv, dpnchd, dpgt1hafh, dppthlt75, dpgt2kmt,
                   pcmcbgv, pcpnchd, pcpgt1hafh, pcppthlt75, pcpgt2kmt,
                   contains('prcp'), contains('tmax')), 
              .funs = list(p1sdp = function(x) x * df$p1sdp,
                           p2shp = function(x) x * df$p2shp,
                           p3rop = function(x) x * df$p3rop)) -> df

df %>%
    select(-date, -key, -n, -cluster, -PRCP, -TMAX, -cases, -deaths, -newcases, -newdeaths) %>%
    colnames() -> cols_to_alterize

stalters <- foreach(i = 1:length(cols_to_alterize), .combine = cbind) %dopar% 
    weightedAlters(df, stateWM, cols_to_alterize[i])

colnames(stalters) <- str_c('stalter_', cols_to_alterize)

alters <- foreach(i = 1:length(cols_to_alterize), .combine = cbind) %dopar% 
    weightedAlters(df, WM, cols_to_alterize[i])

colnames(alters) <- str_c('alter_', cols_to_alterize)

policy_cols <- c('p1sdp', 'p2shp', 'p3rop')

stalters_pn <- foreach(i = 1:length(policy_cols), .combine = cbind) %dopar% 
    weightedAlters(df, stateWM_near, policy_cols[i])

stalters_pf <- foreach(i = 1:length(policy_cols), .combine = cbind) %dopar% 
    weightedAlters(df, stateWM_far, policy_cols[i])

colnames(stalters_pn) <- str_c('stalterNear_', policy_cols)
colnames(stalters_pf) <- str_c('stalterFar_', policy_cols)

df %>%
    bind_cols(as.data.frame(stalters)) %>%
    bind_cols(as.data.frame(alters)) %>%
    bind_cols(as.data.frame(stalters_pn)) %>%
    bind_cols(as.data.frame(stalters_pf)) %>%
    mutate(log_mcbgv    = log(mcbgv),
           log_pnchd    = log(pnchd),
           log_pgt1hafh = log(pgt1hafh),
           log_ppthlt75 = log(ppthlt75),
           log_pgt2kmt  = log(pgt2kmt),
           log_stalter_mcbgv    = log(stalter_mcbgv),
           log_stalter_pnchd    = log(stalter_pnchd),
           log_stalter_pgt1hafh = log(stalter_pgt1hafh),
           log_stalter_ppthlt75 = log(stalter_ppthlt75),
           log_stalter_pgt2kmt  = log(stalter_pgt2kmt)
           log_alter_mcbgv    = log(alter_mcbgv),
           log_alter_pnchd    = log(alter_pnchd),
           log_alter_pgt1hafh = log(alter_pgt1hafh),
           log_alter_ppthlt75 = log(alter_ppthlt75),
           log_alter_pgt2kmt  = log(alter_pgt2kmt)) %>%
    mutate_at(vars(matches('alter')), 
              .funs = list(ep1sdp = function(x) x * df$p1sdp,
                           ep2shp = function(x) x * df$p2shp,
                           ep3rop = function(x) x * df$p3rop)) -> df

df %>%
    bind_cols(as.data.frame(stalters)) %>%
    bind_cols(as.data.frame(alters)) %>%
    bind_cols(as.data.frame(stalters_pn)) %>%
    bind_cols(as.data.frame(stalters_pf)) %>%
    mutate(log_mcbgv    = log(mcbgv),
           log_pnchd    = log(pnchd),
           log_pgt1hafh = log(pgt1hafh),
           log_ppthlt75 = log(ppthlt75),
           log_pgt2kmt  = log(pgt2kmt),
           log_stalter_mcbgv    = log(stalter_mcbgv),
           log_stalter_pnchd    = log(stalter_pnchd),
           log_stalter_pgt1hafh = log(stalter_pgt1hafh),
           log_stalter_ppthlt75 = log(stalter_ppthlt75),
           log_stalter_pgt2kmt  = log(stalter_pgt2kmt),
           log_alter_mcbgv    = log(alter_mcbgv),
           log_alter_pnchd    = log(alter_pnchd),
           log_alter_pgt1hafh = log(alter_pgt1hafh),
           log_alter_ppthlt75 = log(alter_ppthlt75),
           log_alter_pgt2kmt  = log(alter_pgt2kmt)) %>%
    group_by(cluster, date) %>%
    mutate(stnewcases  = sum(newcases),
           stnewdeaths = sum(newdeaths)) -> df

mlist <- str_c(c('January', 'February', 'March', 'April', 'May', 'June', 'July'), ' 2020')
bls_county <- foreach(m = mlist, .combine = rbind) %dopar% get_bls_county(m)
bls_state  <- foreach(m = mlist, .combine = rbind) %dopar% get_bls_state(m)

bls_county %>%
    mutate(mon = month(period)) %>%
    select(key           = fips,
           ct_unemp_rate = unemployed_rate,
           mon) %>%
    inner_join(keys) -> bls_county    

bls_state %>%
    mutate(mon = month(month)) %>%
    select(cluster       = fips_state,
           st_unemp_rate = unemployed_rate,
           mon) -> bls_state

df %>%
    mutate(mon = month(date)) %>%
    left_join(bls_county) %>%
    left_join(bls_state) %>%
    select(-mon) %>%
    ungroup() %>%
    mutate(st_unemp_rate = ifelse(!is.na(st_unemp_rate), st_unemp_rate, ct_unemp_rate),
           mutate(as_newcases  = asinh(newcases),
           as_newdeaths = asinh(newdeaths),
           as_stnewcases  = asinh(stnewcases),
           as_stnewdeaths = asinh(stnewdeaths),
           ct_unemp_rate = as.numeric(ct_unemp_rate),
           st_unemp_rate = as.numeric(st_unemp_rate)) -> df

write_rds(df, str_c(fp, 'PROCESSED_DATA/panel_pre_xgr.RDS'))
