==> data_processing/6a-dml_residualization_main_v2.ipynb <==
library(readr)
library(dplyr)
library(tidyr)
library(stringr)
library(doMC)
library(lfe)
library(xgboost)
registerDoMC(20)

fp <- '/pool001/mfzhao/'
df <- read_rds(str_c(fp, 'PROCESSED_DATA/panel_pre_xgr.RDS'))

df %>%
    select(key,
           date, 
           cluster,
           n,
           log_mcbgv,
           log_pnchd,
           log_pgt1hafh, 
           log_ppthlt75,
           log_pgt2kmt,
           matches('unemp'),
           matches('^[PT][RM][CA][PX]$'),
           matches('^p[123]...$'),
           matches('stalter_p[123]...$'),
           matches('^as_(st)?new')) -> df

colnames(df)

set.seed(2345)
df %>%
    ungroup() %>%
    select(key) %>%
    distinct() %>% 
    mutate(i = sample(1:n(), n(), replace = F),
           fold = i %% 3 + 1) %>%
    select(-i) -> folds

df %>%
    inner_join(folds) %>%
    arrange(date, key) %>%
    ungroup() -> df

df %>%
    ungroup() %>%
    mutate(PRCP.r_fe           = felm(PRCP ~ 0 | key + date, ., weights = df$n)$resid,
           TMAX.r_fe           = felm(TMAX ~ 0 | key + date, ., weights = df$n)$resid,
           as_newcases.r_fe    = felm(as_newcases ~ 0 | key + date, ., weights = df$n)$resid,
           as_newdeaths.r_fe   = felm(as_newdeaths ~ 0 | key + date, ., weights = df$n)$resid,
           as_stnewcases.r_fe  = felm(as_stnewcases ~ 0 | key + date, ., weights = df$n)$resid,
           as_stnewdeaths.r_fe = felm(as_stnewdeaths ~ 0 | key + date, ., weights = df$n)$resid,
           ct_unemp_rate.r_fe  = felm(ct_unemp_rate ~ 0 | key + date, ., weights = df$n)$resid,
           st_unemp_rate.r_fe  = felm(st_unemp_rate ~ 0 | key + date, ., weights = df$n)$resid) %>%
    select(date, 
           key, 
           PRCP.r_fe, 
           TMAX.r_fe, 
           as_newcases.r_fe, 
           as_newdeaths.r_fe, 
           as_stnewcases.r_fe, 
           as_stnewdeaths.r_fe, 
           ct_unemp_rate.r_fe,
           st_unemp_rate.r_fe,
           fold, 
           n) -> residualizer_df

folds <- list(which(folds$fold %in% 1), 
              which(folds$fold %in% 2), 
              which(folds$fold %in% 3))

write_rds(folds, str_c(fp, 'PROCESSED_DATA/folds.RDS'))
write_rds(residualizer_df, str_c(fp, 'PROCESSED_DATA/residualizer_data.RDS'))

XGresidualizer <- function(Y, colname) {
    print(colname)
    residualizer_df %>%
        ungroup() %>% 
        mutate(Y = Y,
               Y.r = felm(Y ~ 0 | key + date, ., weights = residualizer_df$n)$resid) -> temp_df
  
    folds <- list(which(temp_df$fold %in% 1), 
                  which(temp_df$fold %in% 2), 
                  which(temp_df$fold %in% 3))
    
    form <- ~ 0 + PRCP.r_fe + TMAX.r_fe + as_newcases.r_fe + as_newdeaths.r_fe + as_stnewcases.r_fe + as_stnewdeaths.r_fe + ct_unemp_rate.r_fe + st_unemp_rate.r_fe
    
    X     <- model.matrix(form, temp_df)
    dm    <- xgb.DMatrix(data = X, label = temp_df$Y.r)
    param <- list(max_depth=2, eta=.5, silent=1, objective='reg:linear')
    fit   <- xgb.cv(params = param, 
                    data = dm, 
                    folds = folds,
                    nrounds = 100, 
                    early_stopping_rounds = 3, 
                    weight = temp_df$n,
                    nthread = 20)
    best_n <- fit$best_iteration
    for (i in 1:3) {
        tr  <- temp_df %>% filter(fold != i)
        trm <- xgb.DMatrix(data = model.matrix(form, tr), label = tr$Y.r)
        fit <- xgb.train(params = param, 
                         data = trm, 
                         nrounds = best_n, 
                         weight = tr$n,
                         nthread = 20)
        te  <- temp_df %>% filter(fold == i)
        tem <- xgb.DMatrix(data = model.matrix(form, te), label = te$Y.r)
        te %>%
            select(date, key) %>%
            mutate(pred = predict(fit, newdata = tem)) -> pred_df
        assign(str_c('temp',i), pred_df) %>%
        select(-pred, -date, -key)
    }
    out <- bind_rows(temp1, temp2, temp3) %>%
        arrange(date, key) %>%
        mutate(tempname = temp_df$Y.r - pred) %>%
        select(-pred, -date, -key)
    
    colnames(out) <- str_c(colname, '.r')
    return(out)
}

cols_to_xgr <- colnames(
    df %>%
        select(
            matches('^log_'),
            matches('p[123]...')
         )
)
cols_to_xgr

foreach(i = 1:length(cols_to_xgr), .combine = cbind) %do% 
    XGresidualizer(df[[cols_to_xgr[i]]], cols_to_xgr[i]) -> xg.residuals

df %>%
    select(key, date, n, cluster, fold) %>%
    bind_cols(xg.residuals) -> df.r

write_rds(df.r, str_c(fp, 'PROCESSED_DATA/panel_xgr_v3.RDS'))

XGresidualizer <- function(Y, colname) {
    print(colname)
    if(str_detect(colname, 'l[1-7]')) {
        residualizer_df %>%
            mutate(Y = Y) %>%
            group_by(key) %>%
            arrange(key, date) %>%
            filter(row_number() > as.numeric(str_sub(colname, -1, -1))) %>%
            arrange(date, key) %>% 
            ungroup() -> temp_df

        temp_df %>%
            ungroup() %>%
            arrange(date, key) %>%
            mutate(Y.r = felm(Y ~ 0 | key + date, ., weights = temp_df$n)$resid) %>%
            arrange(date, key) -> temp_df
        
    } else {
        residualizer_df %>%
            ungroup() %>% 
            mutate(Y = Y,
                   Y.r = felm(Y ~ 0 | key + date, ., weights = residualizer_df$n)$resid) -> temp_df
    }
    
    folds <- list(which(temp_df$fold %in% 1), 
                  which(temp_df$fold %in% 2), 
                  which(temp_df$fold %in% 3))

    dm    <- xgb.DMatrix(data = model.matrix(~ 0 + PRCP.r_fe + TMAX.r_fe, temp_df), label = temp_df$Y.r)
    param <- list(max_depth=2, eta=.5, silent=1, objective='reg:linear')
    fit   <- xgb.cv(params = param, 
                    data = dm, 
                    folds = folds,
                    nrounds = 100, 
                    early_stopping_rounds = 3, 
                    weight = temp_df$n)
    best_n <- fit$best_iteration
    for (i in 1:3) {
        tr  <- temp_df %>% filter(fold != i)
        trm <- xgb.DMatrix(data = model.matrix(~ 0 + PRCP.r_fe + TMAX.r_fe, tr), label = tr$Y.r)
        fit <- xgb.train(params = param, 
                         data = trm, 
                         nrounds = best_n, 
                         weight = tr$n)
        te  <- temp_df %>% filter(fold == i)
        tem <- xgb.DMatrix(data = model.matrix(~ 0 + PRCP.r_fe + TMAX.r_fe, te), label = te$Y.r)
        te %>%
            select(date, key) %>%
            mutate(pred = predict(fit, newdata = tem)) -> pred_df
        assign(str_c('temp',i), pred_df) %>%
        select(-pred, -date, -key)
    }
    if(str_detect(colname, 'l[1-7]')) {
        out <- residualizer_df %>%
            select(date, key) %>%
            left_join(bind_rows(temp1, temp2, temp3) %>%
                      mutate(tempname = temp_df$Y.r - pred)) %>% 
            arrange(date, key) %>%
            select(-pred, -date, -key)
    } else {
        out <- bind_rows(temp1, temp2, temp3) %>%
            arrange(date, key) %>%
            mutate(tempname = temp_df$Y.r - pred) %>%
            select(-pred, -date, -key)
    }
    colnames(out) <- str_c(colname, '.r')
    return(out)
}
